Several changes and improvements based on Keith's feedback
- Recursive failure definition for CouldNotGetToken

+++ Claire/Client/ApiInfo.swift
@@ -7,11 +7,12 @@
 //
 
 import Foundation
+import UIKit
 
 public struct ApiInfo : Printable
 {
-    var maxImageSize : Int?
-    var minImageSize : Int?
+    var maxImageSize : CGFloat?
+    var minImageSize : CGFloat?
     var maxBatchSize : Int?
     var embedAllowed : Bool?
